+++ client/Packages/com.beamable.server/Editor/BeamServicesCodeWatcher.cs
@@ -374,7 +374,7 @@ namespace Beamable.Server.Editor
 			}
 		}
 
-		public static MicroserviceDescriptor GetGeneratorDescriptor(MicroserviceDescriptor service)
+		public static MicroserviceDescriptor GetGeneratorDescriptor(IDescriptor service)
 		{
 			return new MicroserviceDescriptor
 			{
@@ -384,7 +384,7 @@ namespace Beamable.Server.Editor
 			};
 		}
 
-		public static async Promise StopClientSourceCodeGenerator(MicroserviceDescriptor service)
+		public static async Promise StopClientSourceCodeGenerator(IDescriptor service)
 		{
 			var generatorDesc = GetGeneratorDescriptor(service);
 			var command = new StopImageReturnableCommand(generatorDesc);