+++ client/Packages/com.beamable.server/Editor/UI/Components/MicroserviceVisualElement/CreateMicroserviceVisualElement.cs
@@ -1,6 +1,7 @@
 ﻿using Beamable.Editor.UI.Model;
 using Beamable.Server.Editor;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace Beamable.Editor.Microservice.UI.Components
 {
@@ -8,7 +9,9 @@ namespace Beamable.Editor.Microservice.UI.Components
 	{
 		protected override string NewServiceName { get; set; } = "NewMicroService";
 		protected override string ScriptName => nameof(MicroserviceVisualElement);
-		protected override bool ShouldShowCreateDependentService => MicroservicesDataModel.Instance.Storages.Count != 0;
+
+		protected override bool ShouldShowCreateDependentService =>
+			MicroservicesDataModel.Instance.Storages.Any(x => !x.IsArchived);
 
 		protected override void CreateService(string serviceName, List<ServiceModelBase> additionalReferences = null)
 		{
@@ -17,7 +20,7 @@ namespace Beamable.Editor.Microservice.UI.Components
 		}
 		protected override void InitCreateDependentService()
 		{
-			_serviceCreateDependentService.Init(MicroservicesDataModel.Instance.Storages, "StorageObjects");
+			_serviceCreateDependentService.Init(MicroservicesDataModel.Instance.Storages.Where(x => !x.IsArchived).ToList(), "StorageObjects");
 		}
 	}
 }