+++ client/Packages/com.beamable/CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Beamable button in Unity toolbar should be in correct position for production packages
+- CurrencyHUD no longer throws null reference error when associated currency content has no addressable icon. 
 
 ## [1.2.4]
 no changes